Force unix style line-endings in git.
Checkouts on Windows must preserve unix style line-endings.
(LF instead of CRLF) for provisioning to work.
`/bin/bash` on VM will fail on scripts with CRLF line-endings so
both packer and ansible will fail. These failures are hard to detect
and track down.

On windows we need to use the ansible_local provisioner which will
choke on host_key_checking in `Vagrantfile` so move it to `ansible.cfg`.